third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/read-succeeds.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
const test_desc = 'A read request succeeds and returns the characteristic\'s ' +
    'value.';
const EXPECTED_VALUE = [0, 1, 2];
let characteristic, fake_characteristic;
promise_test(() => getMeasurementIntervalCharacteristic()
  .then(_ => ({characteristic, fake_characteristic} = _))
  .then(() =>
    fake_characteristic.setNextReadResponse(GATT_SUCCESS, EXPECTED_VALUE))
  .then(() => characteristic.readValue())
  .then(value =>
          assert_array_equals(new Uint8Array(value.buffer), EXPECTED_VALUE))
  .then(() => navigator.bluetooth.test.allResponsesConsumed())
  .then(consumed => assert_true(consumed)), test_desc);
</script>
